最小化・最大化・閉じるボタンを新しいComponentに切り出し、macで左、それ以外のOSで右に表示する
