+++ src/components/TitleBarButtons.vue
@@ -36,35 +36,6 @@
         最前面に表示
       </q-tooltip>
     </q-btn>
-    <q-btn
-      dense
-      flat
-      round
-      icon="lens"
-      size="8.5px"
-      color="green"
-      class="title-bar-buttons"
-      @click="maximizeWindow()"
-    ></q-btn>
-    <q-btn
-      dense
-      flat
-      round
-      icon="lens"
-      size="8.5px"
-      color="yellow"
-      class="title-bar-buttons"
-      @click="minimizeWindow()"
-    ></q-btn>
-    <q-btn
-      dense
-      flat
-      icon="lens"
-      size="8.5px"
-      color="red"
-      class="title-bar-buttons"
-      @click="closeWindow()"
-    ></q-btn>
   </q-badge>
   <q-badge
     v-else
@@ -107,72 +78,29 @@
         最前面に表示
       </q-tooltip>
     </q-btn>
-    <q-btn
-      dense
-      flat
-      icon="minimize"
-      class="title-bar-buttons"
-      @click="minimizeWindow()"
-    ></q-btn>
-
-    <q-btn
-      v-if="!isMaximized"
-      dense
-      flat
-      icon="crop_square"
-      class="title-bar-buttons"
-      @click="maximizeWindow()"
-    ></q-btn>
-    <q-btn
-      v-else
-      dense
-      flat
-      :icon="mdiWindowRestore"
-      class="title-bar-buttons"
-      @click="maximizeWindow()"
-    >
-    </q-btn>
-
-    <q-btn
-      dense
-      flat
-      icon="close"
-      class="title-bar-buttons close"
-      @click="closeWindow()"
-    ></q-btn>
   </q-badge>
+  <min-max-close-buttons v-if="!$q.platform.is.mac" />
 </template>
 
 <script lang="ts">
 import { defineComponent, computed } from "vue";
 import { useStore } from "@/store";
-import { mdiWindowRestore } from "@quasar/extras/mdi-v5";
+import MinMaxCloseButtons from "@/components/MinMaxCloseButtons.vue";
 
 export default defineComponent({
   name: "TitleBarButtons",
+  components: { MinMaxCloseButtons },
   setup() {
     const store = useStore();
 
-    const closeWindow = async () => {
-      store.dispatch("CHECK_EDITED_AND_NOT_SAVE");
-    };
-    const minimizeWindow = () => window.electron.minimizeWindow();
-    const maximizeWindow = () => window.electron.maximizeWindow();
     const changePinWindow = () => {
       window.electron.changePinWindow();
     };
 
     const isPinned = computed(() => store.state.isPinned);
 
-    const isMaximized = computed(() => store.state.isMaximized);
-
     return {
-      closeWindow,
-      minimizeWindow,
-      maximizeWindow,
       changePinWindow,
-      mdiWindowRestore,
-      isMaximized,
       isPinned,
     };
   },
@@ -191,9 +119,5 @@ export default defineComponent({
 .title-bar-buttons {
   -webkit-app-region: no-drag;
   overflow: visible;
-}
-
-.close:hover {
-  background-color: red;
 }
 </style>
